event graph, presece graph
