+++ gql_events/GraphTypeDefinitions/PresenceGQLModel.py
@@ -1,0 +1,52 @@
+import strawberry as strawberryA
+import datetime
+from typing import Union, List, Annotated, Optional
+from .utils import getLoaders
+
+
+UserGQLModel = Annotated["UserGQLModel", strawberryA.lazy(".externals")]
+PresenceTypeGQLModel = Annotated["PresenceTypeGQLModel", strawberryA.lazy("..GraphTypeDefinitionsOld")]
+InvitationTypeGQLModel = Annotated["InvitationTypeGQLModel", strawberryA.lazy("..GraphTypeDefinitionsOld")]
+EventGQLModel = Annotated["EventGQLModel", strawberryA.lazy(".EventGQLModel")]
+
+@strawberryA.federation.type(keys=["id"], description="""Describes a relation of an user to the event by invitation (like invited) and participation (like absent)""")
+class PresenceGQLModel:
+
+    @classmethod
+    async def resolve_reference(cls, info: strawberryA.types.Info, id: strawberryA.ID):
+        print(id)
+        if id is None:
+            return None
+        loader = getLoaders(info).presences
+        result = await loader.load(id)
+        if result is not None:
+            result._type_definition = cls._type_definition  # little hack :)
+        return result
+
+    @strawberryA.field(description="""Primary key""")
+    def id(self) -> strawberryA.ID:
+        return self.id
+
+    @strawberryA.field(description="""Time stamp""")
+    def lastchange(self) -> Union[datetime.datetime, None]:
+        return self.lastchange
+
+    @strawberryA.field(description="""Present, Vacation etc.""")
+    async def presence_type(self, info: strawberryA.types.Info) -> Optional["PresenceTypeGQLModel"]:
+        result = await PresenceTypeGQLModel.resolve_reference(info, self.presencetype_id)
+        return result
+
+    @strawberryA.field(description="""Invited, Accepted, etc.""")
+    async def invitation_type(self, info: strawberryA.types.Info) -> Optional["InvitationTypeGQLModel"]:
+        result = await InvitationTypeGQLModel.resolve_reference(info, self.invitation_id)
+        return result
+
+    @strawberryA.field(description="""The user / participant""")
+    def user(self) -> Optional["UserGQLModel"]:
+        result = UserGQLModel(id=self.user_id)
+        return result
+
+    @strawberryA.field(description="""The event""")
+    async def event(self, info: strawberryA.types.Info) -> Optional["EventGQLModel"]:
+        result = await EventGQLModel.resolve_reference(info, id=self.event_id)
+        return result
