+++ gql_events/GraphTypeDefinitions/utils.py
@@ -1,0 +1,13 @@
+from contextlib import asynccontextmanager
+
+@asynccontextmanager
+async def withInfo(info):
+    asyncSessionMaker = info.context["asyncSessionMaker"]
+    async with asyncSessionMaker() as session:
+        try:
+            yield session   
+        finally:
+            pass
+
+def getLoaders(info):
+    return info.context['all']
