+++ main.py
@@ -9,7 +9,7 @@ from strawberry.fastapi import GraphQLRouter
 
 ## Definice GraphQL typu (pomoci strawberry https://strawberry.rocks/)
 ## Strawberry zvoleno kvuli moznosti mit federovane GraphQL API (https://strawberry.rocks/docs/guides/federation, https://www.apollographql.com/docs/federation/)
-from gql_events.GraphTypeDefinitions import Query
+from gql_events.GraphTypeDefinitionsOld import Query
 
 ## Definice DB typu (pomoci SQLAlchemy https://www.sqlalchemy.org/)
 ## SQLAlchemy zvoleno kvuli moznost komunikovat s DB asynchronne
@@ -99,7 +99,7 @@ class MyGraphQL(GraphQL):
         }
 
 
-from gql_events.GraphTypeDefinitions import schema
+from gql_events.GraphTypeDefinitionsOld import schema
 
 ## ASGI app, kterou "moutneme"
 graphql_app = MyGraphQL(schema, graphiql=True, allow_queries_via_get=True)
